Change logger format to json
